% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootPlainFast.R, R/test.R
\name{bootPlainFast}
\alias{bootPlainFast}
\title{Variant 1}
\usage{
bootPlainFast(obj, B = 999, typeRes = "HC3", typeEps = "Liu1988", trunc = 1000)

bootPlainFast(obj, B = 999, typeRes = "HC3", typeEps = "Liu1988", trunc = 1000)
}
\description{
Variant 1

Variant 1
}
