% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootPlainFast2.R
\name{bootPlainFast2}
\alias{bootPlainFast2}
\title{Variant 1}
\usage{
bootPlainFast2(
  obj,
  B = 1000,
  typeRes = "HC3",
  typeEps = "Liu1988",
  trunc = 1000
)
}
\description{
Variables must be names y and x!
}
