% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dcmp.R
\name{dcmp}
\alias{dcmp}
\alias{rcmp}
\title{Conway-Maxwell-Poisson Distribution}
\usage{
dcmp(x, mu, nu, log = FALSE)
}
\arguments{
\item{x}{quantile at which the prob. mass function is evaluated.}

\item{mu}{location parameter}

\item{nu}{dispersion parameter}

\item{log}{Logical. Whether or not the log PMF is returned.}

\item{n}{number of random draws to return}
}
\description{
Probability mass function and random generation for the
Conway-Maxwell-Poisson distribution for given values of the parameters.
}
\details{
Computes PMF and makes random draws from the Conway-Maxwell-Poisson (CMP)
distribution. The PMF of the CMP is given by
\deqn{f(x) = (1/Z) ((nu*mu)^x)/(x!^nu).}
(Guikema and Goffelt 2008).

The normalizing constant is calculated using a combination of finite truncation of the infinite sum
as well as an approximation provided by Shmueli et al. (2005) for small nu and large mu.
}
\examples{
x = rcmp(1000, 10, 2)
plot(table(x)/length(x))
pmf = dcmp(x, 10, 2)
points(x, pmf)
}
\references{
Guikema, Seth D., and Jeremy P. Goffelt. "A flexible count data regression model for risk analysis." Risk analysis 28.1 (2008): 213-223.

Shmueli, Galit, et al. "A useful distribution for fitting discrete data: revival of the Conway–Maxwell–Poisson distribution." Journal of the Royal Statistical Society: Series C (Applied Statistics) 54.1 (2005): 127-142.
}
\author{
Devin S. Johnson and Jeffrey Dunn
}
