% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmpr_helpers.R
\name{cmpr_convert_df_cols_to_query_col_format}
\alias{cmpr_convert_df_cols_to_query_col_format}
\title{Convert the column names of a data frame into a string of values in SQL format for INSERT or UPDATE}
\usage{
cmpr_convert_df_cols_to_query_col_format(df)
}
\arguments{
\item{df}{data frame with column  names to convert}
}
\value{
string of SQL-formatted values for INSERT or UPDATE
}
\description{
Convert the column names of a data frame into a string of values in SQL format for INSERT or UPDATE
}
