% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmpr_helpers.R
\name{cmpr_convert_to_db_cols}
\alias{cmpr_convert_to_db_cols}
\title{Convert from \code{sensorstrings} R column names to CMPDB column names}
\usage{
cmpr_convert_to_db_cols(df)
}
\arguments{
\item{df}{data frame with columns to be renamed}
}
\value{
columns
}
\description{
Convert from \code{sensorstrings} R column names to CMPDB column names
}
