% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmpr_helpers.R
\name{cmpr_convert_to_ss_cols}
\alias{cmpr_convert_to_ss_cols}
\title{Convert from CMPDB column names to \code{sensorstrings} R package columns}
\usage{
cmpr_convert_to_ss_cols(df)
}
\arguments{
\item{df}{data frame with columns to be renamed.}
}
\value{
columns
}
\description{
Convert from CMPDB column names to \code{sensorstrings} R package columns
}
