% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmpr_filter_depl_metadata.R
\name{cmpr_filter_depl_metadata}
\alias{cmpr_filter_depl_metadata}
\title{Filter Deployment Metadata by Date}
\usage{
cmpr_filter_depl_metadata(metadata_df, last_update_date = NULL)
}
\arguments{
\item{metadata_df}{data frame of sensor string deployment metadata. Must
include column \code{last_updated_date}.}

\item{last_update_date}{date when the database was last updated from the
metadata tracking sheet, in yyyy-mm-dd format.}
}
\value{
Returns rows of \code{metadata_df} that were updated after
\code{last_update_date}.
}
\description{
Filter Deployment Metadata by Date
}
