% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmpr_get_county_metadata.R
\name{cmpr_get_county_metadata}
\alias{cmpr_get_county_metadata}
\title{Get County Metadata}
\usage{
cmpr_get_county_metadata(conn)
}
\arguments{
\item{conn}{database connection object}
}
\value{
data frame with county metadata
}
\description{
Get County Metadata
}
