% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmpr_helpers.R
\name{cmpr_get_ss_cmpdb_col_mapping}
\alias{cmpr_get_ss_cmpdb_col_mapping}
\title{Get a mapping of \code{sensorstrings} column names to CMPDB column names}
\usage{
cmpr_get_ss_cmpdb_col_mapping(mode)
}
\arguments{
\item{mode}{string "forward" or "reverse" to indicate whether to return a
mapping of \code{sensorstrings} columns names to CMPDB column names or CMPDB
column names to \code{sensorstrings} column names.}
}
\value{
named vector which maps each \code{sensorstrings} column name to each
CMPDB column name or vice versa. Only returns a mapping for column names
which differ between \code{sensorstrings} and CMPDB.
}
\description{
Get a mapping of \code{sensorstrings} column names to CMPDB column names
}
