% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmpr_get_station_data.R
\name{cmpr_get_station_data}
\alias{cmpr_get_station_data}
\title{Get list of CMP stations}
\usage{
cmpr_get_station_data(conn)
}
\arguments{
\item{conn}{database connection object}
}
\value{
data frame of all stations and their attributes
}
\description{
Get list of CMP stations
}
