% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmpr_get_station_depl.R
\name{cmpr_get_station_depl}
\alias{cmpr_get_station_depl}
\title{Get list of deployments at a given station}
\usage{
cmpr_get_station_depl(conn, station_name)
}
\arguments{
\item{conn}{database connection object.}

\item{station_name}{name of station}
}
\value{
data frame of all of the deployments from a station
}
\description{
Get list of deployments at a given station
}
