% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmpr_get_waterbody_metadata.R
\name{cmpr_get_waterbody_metadata}
\alias{cmpr_get_waterbody_metadata}
\title{Get Waterbody Metadata}
\usage{
cmpr_get_waterbody_metadata(conn)
}
\arguments{
\item{conn}{database connection object}
}
\value{
data frame with waterbody metadata
}
\description{
Get Waterbody Metadata
}
