% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmpr_import_depl_metadata.R
\name{cmpr_import_depl_metadata}
\alias{cmpr_import_depl_metadata}
\title{Import Deployment Metadata}
\usage{
cmpr_import_depl_metadata(filepath)
}
\arguments{
\item{filepath}{location of the metadata tracking sheet Excel file, include
file name and extension.}
}
\value{
data frame of the deployment metadata sheet
}
\description{
Import Deployment Metadata
}
