% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmpr_import_metadata.R
\name{cmpr_import_metadata}
\alias{cmpr_import_metadata}
\title{Import Metadata to DB}
\usage{
cmpr_import_metadata(conn, filepath)
}
\arguments{
\item{conn}{database connection object}

\item{filepath}{location of the metadata tracking sheet Excel file}
}
\value{
tbd
}
\description{
Import Metadata to DB
}
