% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmpr_insert_location_metadata.R
\name{cmpr_insert_location_metadata}
\alias{cmpr_insert_location_metadata}
\title{Title}
\usage{
cmpr_insert_location_metadata(values, mode = "station")
}
\arguments{
\item{mode}{}
}
\description{
Title
}
