% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmpr_helpers.R
\name{cmpr_prepend_lease_zeroes}
\alias{cmpr_prepend_lease_zeroes}
\title{Prepend any missing zeroes to lease numbers}
\usage{
cmpr_prepend_lease_zeroes(lease_num)
}
\arguments{
\item{lease_num}{lease number that may be missing prepended zeroes due to spreadsheet formatting idiosyncracies}
}
\value{
string lease number with prepended zeroes
}
\description{
Prepend any missing zeroes to lease numbers
}
