% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmpr_update_db_depl_metadata.R
\name{cmpr_update_db_depl_metadata}
\alias{cmpr_update_db_depl_metadata}
\title{Update CMP database with new deployment metadata}
\usage{
cmpr_update_db_depl_metadata(conn, metadata_sheet)
}
\arguments{
\item{conn}{database connection object}

\item{metadata_sheet}{data frame of cleaned data from the CMP metadata tracking sheet}
}
\value{
tbd
}
\description{
Update CMP database with new deployment metadata
}
