% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmpr_update_location_metadata.R
\name{cmpr_update_location_metadata}
\alias{cmpr_update_location_metadata}
\title{Update Location Metadata in database}
\usage{
cmpr_update_location_metadata(conn, updated_location_metadata, mode)
}
\arguments{
\item{conn}{database connection object}

\item{updated_location_metadata}{data frame of updated location metadata, matching database format,}

\item{mode}{indicates which kind of location data is coming into the database and thus which tables it
should go into, options are 'waterbody' and 'station'}
}
\value{
tbd
}
\description{
Update Location Metadata in database
}
