% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_to_db.R
\name{connect_to_db}
\alias{connect_to_db}
\title{Connect to CMP Database}
\usage{
connect_to_db(connection_config = "default")
}
\arguments{
\item{connection_config}{name of configuration in config.yml to use for connection}
}
\value{
database connection object
}
\description{
Connect to CMP Database
}
\examples{
db_conn <- try(connect_to_db())
}
