% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_station_list.R
\name{get_station_list}
\alias{get_station_list}
\title{Get list of CMP stations}
\usage{
get_station_list(conn)
}
\arguments{
\item{conn}{database connection object}
}
\value{
data frame of all stations and their attributes
}
\description{
Get list of CMP stations
}
