% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_depl_metadata.R
\name{import_depl_metadata}
\alias{import_depl_metadata}
\title{Import Deployment Metadata}
\usage{
import_depl_metadata(filepath)
}
\arguments{
\item{filepath}{location of the metadata tracking sheet Excel file}
}
\value{
data.frame metadata sheet
}
\description{
Import Deployment Metadata
}
\examples{
try(
  import_depl_metadata("./extdata/water_quality_deployment_tracking.xlsx")
)
}
