% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmpR.R
\name{initialize,cmpR-method}
\alias{initialize,cmpR-method}
\title{Constructor cmpR}
\usage{
\S4method{initialize}{cmpR}(.Object, data = data.frame, formula = formula)
}
\arguments{
\item{.Object}{cmpR object}

\item{data}{data.frame with data to be evalauted}
}
\description{
Constructor for new instance
}
