% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tcalFast.R
\name{tcalFast2}
\alias{tcalFast2}
\title{Calculate t-value, fast implementation}
\usage{
tcalFast2(obj, rep = 100, method = "V2", res = "Liu1988", eps = "HC3", ...)
}
\description{
data must adhere to the form y~x
}
