% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmpr_helpers.R
\name{cmpr_clean_enum_types}
\alias{cmpr_clean_enum_types}
\title{Clean up PostgreSQL custom enum type columns}
\usage{
cmpr_clean_enum_types(df)
}
\arguments{
\item{df}{data frame with columns to clean.}
}
\value{
data frame with column types converted to R types
}
\description{
Clean up PostgreSQL custom enum type columns
}
