% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmpr_helpers.R
\name{cmpr_convert_df_to_query_value_format}
\alias{cmpr_convert_df_to_query_value_format}
\title{Convert a data frame into a string of values in SQL format for INSERT or UPDATE}
\usage{
cmpr_convert_df_to_query_value_format(df)
}
\arguments{
\item{df}{data frame to convert}
}
\value{
string of SQL-formatted values for INSERT or UPDATE operations
}
\description{
Convert a data frame into a string of values in SQL format for INSERT or UPDATE
}
