% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmpr_get_station_data.R
\name{cmpr_get_station_data}
\alias{cmpr_get_station_data}
\title{Get measurement data from a specific CMP deployment}
\usage{
cmpr_get_station_data(
  conn,
  station_name,
  variable_type = NULL,
  variable_name = NULL
)
}
\arguments{
\item{conn}{database connection object.}

\item{station_name}{character vectors of the name of the station(s) of interest.}

\item{variable_type}{type of variable to extract. Must be one of
"dissolved_oxygen", "salinity", "sensor_depth_measured", or "temperature".
This is useful for extracting dissolved oxygen data measured in both units.
If \code{variable_type} is specified, \code{variable_name} must be
\code{NULL}.}

\item{variable_name}{name of variable to extract. Must be one of:
"dissolved_oxygen_percent_saturation", "dissolved_oxygen_mg_per_l",
"salinity_psu", "sensor_depth_measured_m", or "temperature_degree_c". If
\code{variable_name} is specified, \code{variable_type} must be
\code{NULL}.}
}
\value{
data frame of the specified variable measured at the specified
station(s).
}
\description{
Get measurement data from a specific CMP deployment
}
