% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmpr_get_station_data.R
\name{cmpr_get_station_data}
\alias{cmpr_get_station_data}
\title{Get measurement data from a specific CMP deployment}
\usage{
cmpr_get_station_data(
  conn,
  station_name,
  variable_type = NULL,
  variable_name = NULL
)
}
\arguments{
\item{conn}{database connection object}

\item{station_name}{name of the station(s)}

\item{variable_type}{type of variable to filter the data for, excludes use of variable_name}

\item{variable_name}{name of variable to filter the data for, excludes use of variable_type}
}
\value{
data frame of data measured over the course of a deployment
}
\description{
Get measurement data from a specific CMP deployment
}
