% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmpr_import_depl_metadata_sheet.R
\name{cmpr_import_depl_metadata_sheet}
\alias{cmpr_import_depl_metadata_sheet}
\title{Import Deployment Metadata}
\usage{
cmpr_import_depl_metadata_sheet(filepath)
}
\arguments{
\item{filepath}{location of the metadata tracking sheet Excel file}
}
\value{
data.frame metadata sheet
}
\description{
Import Deployment Metadata
}
