% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmpr_import_location_metadata.R
\name{cmpr_import_location_metadata}
\alias{cmpr_import_location_metadata}
\title{Import Location Metadata}
\usage{
cmpr_import_location_metadata(conn, location_metadata_sheet)
}
\arguments{
\item{conn}{database connection object}

\item{location_metadata_sheet}{data frame of values read in from the location metadata Excel sheetS}
}
\value{
tbd
}
\description{
Import Location Metadata
}
