% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmpr_insert_location_metadata.R
\name{cmpr_insert_location_metadata}
\alias{cmpr_insert_location_metadata}
\title{Insert Location Metadata into database}
\usage{
cmpr_insert_location_metadata(conn, new_location_metadata, mode)
}
\arguments{
\item{conn}{database connection object}

\item{new_location_metadata}{data frame of new location metadata, matching database format,}

\item{mode}{indicates which kind of location data is coming into the database and thus which tables it
should go into, options are 'waterbody' and 'station'}
}
\value{
tbd
}
\description{
Insert Location Metadata into database
}
