% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmpr_helpers.R
\name{cmpr_list_enum_columns}
\alias{cmpr_list_enum_columns}
\title{Columns in the CMPDB that are custom PostgreSQL types}
\usage{
cmpr_list_enum_columns()
}
\value{
vector of strings listing columns in the CMPDB that are custom
PostgreSQL types.
}
\description{
Columns in the CMPDB that are custom PostgreSQL types
}
