% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmpr_record_data_import.R
\name{cmpr_record_data_import}
\alias{cmpr_record_data_import}
\title{Write a new data import record and submit it to the CMPDB}
\usage{
cmpr_record_data_import(conn, data_name, notes = "")
}
\arguments{
\item{conn}{}
}
\description{
Write a new data import record and submit it to the CMPDB
}
