% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmpr_update_depl_metadata.R
\name{cmpr_update_depl_metadata}
\alias{cmpr_update_depl_metadata}
\title{Update Deployment Metadata}
\usage{
cmpr_update_depl_metadata(conn, filepath)
}
\arguments{
\item{conn}{database connection object}

\item{filepath}{location of the metadata tracking sheet Excel file}
}
\description{
Update Deployment Metadata
}
