% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmpr_validate_metadata_sheet.R
\name{cmpr_validate_metadata_sheet}
\alias{cmpr_validate_metadata_sheet}
\title{Validate metadata sheet}
\usage{
cmpr_validate_metadata_sheet(conn, metadata_sheet)
}
\arguments{
\item{conn}{database connection object}

\item{metadata_sheet}{data frame of metadata pulled from the CMP metadata tracking sheet}
}
\value{
NA
}
\description{
Validate metadata sheet
}
