% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{parse_time_from_excel}
\alias{parse_time_from_excel}
\title{Parse time from date time}
\usage{
parse_time_from_excel(x)
}
\arguments{
\item{x}{datetime string in YYYY-MM-DD HH:MM:SS format}
}
\value{
string representing a time in the format HH:MM:SS
}
\description{
Parse time from date time
}
\examples{
parse_time_from_excel("2024-10-10 10:00:00")
}
