% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{parse_time_from_excel}
\alias{parse_time_from_excel}
\title{Parse time from POSIXct}
\usage{
parse_time_from_excel(x)
}
\arguments{
\item{x}{POSIXct value or vector of POSIXct values}
}
\value{
string representing a time in the format HH:MM:SS
}
\description{
Parse time from POSIXct
}
\examples{
datetime_from_excel <- as.POSIXct("2024-10-10 10:00:00")
parse_time_from_excel(datetime_from_excel)
}
